StageProjet/requirements.txt
Python==3.8.20
cassandra-driver== 3.29.2
Flask==2.2.5
joblib==1.4.2
requests==2.32.3
requests-toolbelt==1.0.0
resume-parser==0.8.4
scikit-learn==1.3.2
scipy==1.10.1
selenium==4.26.1
numpy==1.24.4
nltk==3.9.1
lxml==5.3.0
pandas==2.0.3
ftfy==6.2.3